+++ src/test/java/calculator/string/StringSplitterTest.java
@@ -5,8 +5,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import caculator.string.StringSplitter;
-
 class StringSplitterTest {
 
     @Test
