+++ src/test/java/calculator/string/StringTest.java
@@ -6,8 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import caculator.string.StringSupport;
-
 class StringTest {
 
     public static final String DELIMITER = ",";
